Greedy Algorithm

+++ Algoritmos/factorialRecursivo.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int factorial(int n){
+long long int factorial(int n){
     if(n == 0){
         return (1);
     }
